+++ projects/enneagone-angular-ds/src/lib/button/button.component.ts
@@ -1,4 +1,4 @@
-import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
+import {Component, EventEmitter, Input, Output} from '@angular/core';
 
 @Component({
   selector: 'e9-button',
@@ -8,13 +8,14 @@ import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
 export class ButtonComponent {
   @Input() value: string;
   @Input() color: string;
+  @Input() textColor: string;
   @Input() disabled: boolean;
 
   @Output() btnClick: EventEmitter<any> = new EventEmitter();
-
   constructor() {
     this.disabled = false;
     // TODO: Set up les couleurs par defaut sur la lib
     this.color = '#7C72A0';
+    this.textColor = '#FFFFFF';
   }
 }